+++ crosswrd-app/src/EditLights.tsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
 import { Container } from "react-bootstrap";
 
-import { Grid, Lights } from "./Lights";
-import { displayGrid, renderCells } from "./Grid";
+import { Lights } from "./Lights";
+import { Grid, displayGrid, renderCells } from "./Grid";
 import { WrappedRow, StateSetter, ToggleButton } from "./Helpers";
 import { ExportLights } from "./ExportLights";
