+++ crosswrd-app/src/Grid.tsx
@@ -1,4 +1,5 @@
 import { OrderedMap } from "immutable";
+import styled from "styled-components";
 
 import { Cell, CellProps } from "./Cell";
 import { Lights, toggleCell } from "./Lights";
@@ -36,3 +37,39 @@ export const renderCells = (
       toggleCell={toggleCell(setLights, r)}
     />
   ));
+
+type RawGridProps = {
+  size: number;
+  cellSize: string;
+};
+
+const RawGrid = styled.div<RawGridProps>`
+  padding: 1px;
+  display: grid;
+  place-content: start center;
+  grid-template-columns: repeat(
+    ${(props) => props.size},
+    ${(props) => props.cellSize}
+  );
+  grid-template-columns: repeat(
+    ${(props) => props.size},
+    ${(props) => props.cellSize}
+  );
+  border: 1px black;
+  width: fit-content;
+  background-color: black;
+  column-gap: 1px;
+  row-gap: 1px;
+  margin: auto;
+`;
+
+export type GridProps = {
+  size: bigint;
+  children: JSX.Element[];
+};
+
+export const Grid = ({ size, children }: GridProps) => (
+  <RawGrid size={Number(size)} cellSize="30px">
+    {children}
+  </RawGrid>
+);
